fish_api/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
uvloop==0.19.0
websockets==12.0
pillow==10.2.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
matplotlib==3.8.2
requests==2.31.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1
shapely==2.0.2